Squashed commit of the following:

commit 0bd45d2f9d8311efa8d1c12dff81c78a067961ed
Merge: 91f0a14 fa0ee7a
Author: Pawel Nowak <pawel.nowak@relativity.com>
Date:   Thu Apr 19 14:44:57 2018 +0200

    Merge branch 'develop' into merge-from-fer

commit 91f0a1447a65be00afa62e51289cce9f8156b93c
Author: Pawel Nowak <pawel.nowak@relativity.com>
Date:   Thu Apr 19 11:33:47 2018 +0200

    REL-218036 Changes in tests to fix build error

commit 743df7dd78be60446b32c9bb834b467f945ecbd3
Merge: 86d4d61 0ad5fc3
Author: Pawel Nowak <pawel.nowak@relativity.com>
Date:   Wed Apr 18 15:41:55 2018 +0200

    Merge branch 'release-9.6-february-4-hot-fix' into merge-from-fer

commit 0ad5fc3da2fa74055dd720cc5132fb551e697695
Author: Patryk Spytkowski <patryk.spytkowski@relativity.com>
Date:   Fri Mar 2 07:22:19 2018 -0600

    Merge pull request #626 in IN/integrationpoints from release-9.6-february-REL-204625 to release-9.6-february-4-hot-fix

    Squashed commit of the following:

    commit 82a978eb9446765edd412fa3777327a77d47aa50
    Merge: 072f5e9 26c8289
    Author: Patryk Spytkowski <patryk.spytkowski@relativity.com>
    Date:   Fri Mar 2 14:18:42 2018 +0100

        Merge branch 'release-9.6-february-4-hot-fix' into release-9.6-february-REL-204625

    commit 072f5e938af3843a528f1d4d8613ee5aaa358533
    Author: Patryk Spytkowski <patryk.spytkowski@relativity.com>
    Date:   Fri Mar 2 13:33:08 2018 +0100

        Removed unit tests

    commit 693bdaf1832bb93e2d62aee46ed67f91f142b800
    Author: Patryk Spytkowski <patryk.spytkowski@relativity.com>
    Date:   Fri Mar 2 12:42:06 2018 +0100

        Fix

commit 26c82898c36f11083f636acb29ad4daaf8aaf723
Author: Andrzej Korczyński <andrzej.korczynski@relativity.com>
Date:   Fri Mar 2 06:03:00 2018 -0600

    Merge pull request #625 in IN/integrationpoints from release-9.6-february-REL-204625 to release-9.6-february-4-hot-fix

    Squashed commit of the following:

    commit 693bdaf1832bb93e2d62aee46ed67f91f142b800
    Author: Patryk Spytkowski <patryk.spytkowski@relativity.com>
    Date:   Fri Mar 2 12:42:06 2018 +0100

        Fix

commit a2cfc8225fe880cfa62b12401590f6a2c5179e93
Author: Patryk Spytkowski <patryk.spytkowski@relativity.com>
Date:   Fri Mar 2 10:39:34 2018 +0100

    Disabled rethrowing of an exception JobStatistics

commit ecef0e87ba3724c4376d7911ba2bc8c476a4afeb
Author: Artur Mitianiec <artur.mitianiec@relativity.com>
Date:   Tue Feb 20 08:19:59 2018 -0600

    Merge pull request #594 in IN/integrationpoints from REL-201888-r1-incident-integration-points to release-9.6-april

    Squashed commit of the following:

    commit fac1f35b403b72b0a37c011e930d7a24d2ebe60c
    Author: Michał Grygierzec <michal.grygierzec@relativity.com>
    Date:   Tue Feb 20 14:56:15 2018 +0100

        Fix Unit Tests for NativeFileSizeStatistics

    commit e6c1519d55a630adbb35ad6229ca8b55cc06047d
    Author: Michał Grygierzec <michal.grygierzec@relativity.com>
    Date:   Tue Feb 20 14:54:19 2018 +0100

        [REL-201888] Fix Integration points error Job status failed

    (cherry picked from commit adb81c1398134bfbbdf96f6ceac69bdf118d5a63)

commit 0ad36b4dd2a339ba189ec8d9fbd34b320c6e5162
Author: Artur Mitianiec <artur.mitianiec@relativity.com>
Date:   Tue Feb 6 07:28:07 2018 -0600

    Merge pull request #545 in IN/integrationpoints from REL-197209 to release-9.6-february-2-hot-fix

    Squashed commit of the following:

    commit 84af45155b5ef36f5d474a26fd66808013d11ac5
    Author: Patryk Spytkowski <patryk.spytkowski@relativity.com>
    Date:   Tue Feb 6 14:24:31 2018 +0100

        Added null as ToggleProvider

    commit c200fe3b73a618718cebd5d5cab8e4b781091165
    Author: Patryk Spytkowski <patryk.spytkowski@relativity.com>
    Date:   Tue Feb 6 14:22:54 2018 +0100

        changed file name

    commit 15f546e37c5412367992918801a1424564f269d6
    Author: Patryk Spytkowski <patryk.spytkowski@relativity.com>
    Date:   Tue Feb 6 12:46:08 2018 +0100

        Added toggle disabling DG optimization instead of earlier attempts

    commit 405df8b71bd888bc85de5c38cfdd868bad27c5f6
    Author: Patryk Spytkowski <patryk.spytkowski@relativity.com>
    Date:   Tue Feb 6 12:20:45 2018 +0100

        Changes in setting flag for DG optimization

+++ Source/kCura.IntegrationPoints.Data/Statistics/Implementations/NativeFileSizeStatistics.cs
@@ -42,7 +42,7 @@ namespace kCura.IntegrationPoints.Data.Statistics.Implementations
 			catch (Exception e)
 			{
 				_logger.LogError(e, _FOR_FOLDER_ERROR, folderId, viewId);
-			    return 0;
+				return 0;
 			}
 		}
 
@@ -61,7 +61,7 @@ namespace kCura.IntegrationPoints.Data.Statistics.Implementations
 			catch (Exception e)
 			{
 				_logger.LogError(e, _FOR_PRODUCTION_ERROR, productionSetId);
-			    return 0;
+				return 0;
 			}
 		}
 
@@ -83,7 +83,7 @@ namespace kCura.IntegrationPoints.Data.Statistics.Implementations
 			catch (Exception e)
 			{
 				_logger.LogError(e, _FOR_SAVED_SEARCH_ERROR, savedSearchId);
-			    return 0;
+				return 0;
 			}
 		}
 
@@ -97,7 +97,7 @@ namespace kCura.IntegrationPoints.Data.Statistics.Implementations
 
 		private int GetArtifactId(RelativityObject relativityObject) => relativityObject.ArtifactID;
 
-		private int GetTotalFileSize(IEnumerable<int> artifactIds, int workspaceArtifactId)
+		private long GetTotalFileSize(IEnumerable<int> artifactIds, int workspaceArtifactId)
 		{
 			const string sqlText = "SELECT COALESCE(SUM([Size]),0) FROM [File] WHERE [Type] = @FileType AND [DocumentArtifactID] IN (SELECT * FROM @ArtifactIds)";
 
@@ -113,7 +113,7 @@ namespace kCura.IntegrationPoints.Data.Statistics.Implementations
 			};
 
 			IDBContext dbContext = _helper.GetDBContext(workspaceArtifactId);
-			return dbContext.ExecuteSqlStatementAsScalar<int>(sqlText, artifactIdsParameter, fileTypeParameter);
+			return dbContext.ExecuteSqlStatementAsScalar<long>(sqlText, artifactIdsParameter, fileTypeParameter);
 		}
 
 	}
